FaultService Fault entity,request and response added updateDateTime

+++ emobility-hub/src/main/java/org/unibl/etf/emobility_hub/repositories/FaultEntityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import org.unibl.etf.emobility_hub.models.domain.entity.FaultEntity;
 
 @Repository
-public interface FaultEntityRepository extends JpaRepository<FaultEntity, Integer> {
+public interface FaultEntityRepository extends JpaRepository<FaultEntity, Long> {
 }
